fix(tasks): move query parser behind facade
Keep the query endpoint within the Tasks product's presentation-to-facade import boundary.

Generated-By: PostHog Desktop
Task-Id: 88cebdfa-522e-41f6-8b32-ba598dcf13ef
